[!24][COMPATIBILITY] Improving compatibility with different types of bert-based models
# Why is the change needed?

Currently, the classes used by transformers library are all associated to the Bert architecture (e.g., BertTokenizer). Then they are not compatible with other types of pretrained models, like Roberta, and this results in errors like

```
The tokenizer class you load from this checkpoint is not the same type as the class this function is called from. It may result in unexpected tokenization.
The tokenizer class you load from this checkpoint is 'RobertaTokenizer'.
The class this function is called from is 'BertTokenizer'.
```

# What changes does the patch introduce?

This modification adds the use Auto Classes instead of the specific classes associated to single types of model (e.g., roberta, camembert, bert, and so on).
Auto Classes automatically understand which type of model (or tokenizer) you are loading based on the config file associated with each model and on the name of the model (or tokenizer) you are loading.For further information on Auto Classes, see https://huggingface.co/docs/transformers/main/model_doc/auto

Moreover, the name of some classes are changed to more generic names (e.g., BertPreprocessor -> Preprocessor), as the repository is intended to work with different types of Bert-based architectures, not only with the vanilla Bert architecture.

# How was this patch tested?

With this modification, training and generate with different types of Bert work.