[!28][EVAL] Adding neomorphemes evaluation script
# Why is the change needed?

This change introduces the evaluation script for Neo-GATE and the corresponding unit test script and resources in the repository.

# What changes does the patch introduce?

The patch adds the evaluation script `neomorphemes_evaluation.py` in `src`.

It also adds the unit test script `neomorphemes_unit_tests.py`, as well as the resources associated to it included in `neomorphemes_resources`, in the `test` folder.

# How was this patch tested?

Unit tests and manual checks with desired outputs.