[!33][MISC] Improve and update README.md and GeNTE.md
# Why is the change needed?

Repository documents need some maintenance as some of the bib citations and links for one article are now outdated since they referred to a pre-print version of an article while the official version from the proceedings is available on the ACL Anthology is available.
Moreover, there are small errors that need to be fixed (missing info and typos).

# What changes does the patch introduce?

With this change we update repository documents, include missing information, and correct a few typos and mistakes.

# How was this patch tested?

No test needed.